:recycle: Refactor path wasm code to its own wasm submodule
