+++ render-wasm/src/wasm.rs
@@ -1,4 +1,5 @@
 pub mod fills;
 pub mod fonts;
+pub mod paths;
 pub mod strokes;
 pub mod text;
